EP-790 Moved generateRandomUniqueValue to EntityManager
